agent/pyproject.toml
[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.backends._legacy:_Backend"

[project]
name = "clapcheeks"
version = "0.1.0"
description = "Clap Cheeks — AI Dating Co-Pilot local agent"
requires-python = ">=3.9"
dependencies = [
    "click>=8.1",
    "rich>=13.0",
    "requests>=2.31",
    "pyyaml>=6.0",
    "keyring>=24.0",
]

[project.scripts]
clapcheeks = "clapcheeks.cli:main"
